+++ packages/react/src/experimental/Information/Avatars/PersonAvatar/index.stories.tsx
@@ -1,28 +1,52 @@
 import type { Meta, StoryObj } from "@storybook/react"
 import { Check } from "../../../../icons/app"
 import { sizes } from "../../../../ui/avatar"
-import { PersonAvatar } from "./index"
+import { PersonAvatar, PersonAvatarProps } from "./index"
 
-const meta: Meta<typeof PersonAvatar> = {
-  component: PersonAvatar,
+const PersonAvatarExample = (
+  props: PersonAvatarProps & { hasBadge: boolean }
+) => {
+  return <PersonAvatar {...props} />
+}
+
+const meta: Meta<typeof PersonAvatarExample> = {
+  component: PersonAvatarExample,
   title: "Avatars/PersonAvatar",
   tags: ["autodocs", "experimental"],
   argTypes: {
     size: {
       control: "select",
       options: sizes,
+      description: "Select the size of the avatar",
+    },
+    hasBadge: {
+      control: "boolean",
+      description: "Toggle badge visibility",
+    },
+    badge: {
+      table: { disable: true },
     },
   },
   args: {
     firstName: "Dani",
     lastName: "Moreno",
     size: "medium",
+    hasBadge: false,
   },
-} satisfies Meta<typeof PersonAvatar>
+  decorators: [
+    (Story, context) => {
+      const { hasBadge, ...args } = context.args
+      const badgeProps = hasBadge
+        ? { type: "positive" as const, icon: Check }
+        : undefined
+      return <Story args={{ ...args, badge: badgeProps }} />
+    },
+  ],
+} satisfies Meta<typeof PersonAvatarExample>
 
 export default meta
 
-type Story = StoryObj<typeof PersonAvatar>
+type Story = StoryObj<typeof PersonAvatarExample>
 
 export const Default: Story = {}
 
@@ -38,5 +62,6 @@ export const WithBadge: Story = {
       type: "positive",
       icon: Check,
     },
+    size: "medium",
   },
 }