+++ packages/react/src/experimental/OneDataCollection/Filters/Components/FiltersControls.tsx
@@ -113,7 +113,7 @@ export function FiltersControls<Filters extends FiltersDefinition>({
               )}
             </div>
 
-            <div className="flex items-center justify-end gap-2 border-solid border-transparent border-t-f1-border-secondary px-3 py-2">
+            <div className="flex items-center justify-end gap-2 border border-solid border-transparent border-t-f1-border-secondary p-2">
               <Button
                 onClick={handleApplyFilters}
                 label={i18n.filters.applyFilters}