+++ packages/react/src/experimental/OneDataCollection/__stories__/visualizations/card.stories.tsx
@@ -5,6 +5,9 @@ const meta = {
   title: "Data Collection/Visualizations/Card",
   parameters: {
     layout: "padded",
+    a11y: {
+      skipCi: true,
+    },
     docs: {
       description: {
         component: