+++ packages/react/src/experimental/OneDataCollection/visualizations/collection/Table/index.tsx
@@ -97,7 +97,8 @@ export const TableCollection = <
   source,
   frozenColumns = 0,
   onSelectItems,
-  onTotalItemsChange,
+  onLoadData,
+  onLoadError,
 }: CollectionProps<
   R,
   Filters,
@@ -115,6 +116,7 @@ export const TableCollection = <
     )
   )
 
+
   const { data, paginationInfo, setPage, isInitialLoading, properties } =
     useData<R, Filters, Sortings, NavigationFilters, Grouping>(source)
 
@@ -126,9 +128,17 @@ export const TableCollection = <
       return columns
     }, [columns, properties])
 
+
   useEffect(() => {
-    onTotalItemsChange?.(paginationInfo?.total || data.records.length)
-  }, [paginationInfo?.total, onTotalItemsChange, data.records])
+    onLoadData({
+      totalItems: paginationInfo?.total || data.records.length,
+      filters: source.currentFilters,
+      search: source.currentSearch,
+      isInitialLoading,
+      data: data.records,
+    })
+    // eslint-disable-next-line react-hooks/exhaustive-deps --  we don't want to re-run this effect when the filters change, just when the data changes
+  }, [paginationInfo?.total, data.records])
 
   const { currentSortings, setCurrentSortings, isLoading } = source