+++ packages/core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.12.2](https://github.com/factorialco/factorial-one/compare/factorial-one-core-v1.12.1...factorial-one-core-v1.12.2) (2025-05-30)
+
+
+### Bug Fixes
+
+* change upsell icon color ([#1943](https://github.com/factorialco/factorial-one/issues/1943)) ([0d5a1b1](https://github.com/factorialco/factorial-one/commit/0d5a1b17b894281a9118b68a658361c9c71f11cf))
+
+## [1.12.1](https://github.com/factorialco/factorial-one/compare/factorial-one-core-v1.12.0...factorial-one-core-v1.12.1) (2025-05-29)
+
+
+### Bug Fixes
+
+* apply dark mode to our components ([#1916](https://github.com/factorialco/factorial-one/issues/1916)) ([488ea14](https://github.com/factorialco/factorial-one/commit/488ea146128980a77323e24babbdd8a04b174c3f))
+
 ## [1.12.0](https://github.com/factorialco/factorial-one/compare/factorial-one-core-v1.11.0...factorial-one-core-v1.12.0) (2025-05-23)